requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
mangum
openai
boto3
pillow
pyjwt
